requirements.txt
Flask
pytesseract
Pillow
spacy
pyyaml
pdfplumber
